IND-524 added functional tests, moved unit tests
